requirements.txt
streamlit
pymupdf
google-generativeai
python-dotenv
chromadb==0.3.23
pydantic<2